selectors/learned-selector/pyproject.toml
[project]
name = "Control Full Reaction Learned Selector"
version = "0.1.0"
description = "Learned Selector for CSTR"
authors = [{ name = "John Doe", email = "john.doe@composabl.com" }]
dependencies = [
    "composabl-core"
]

[composabl]
type = "selector-teacher"
entrypoint = "learned_selector.teacher:Teacher"
